+++ ofdrw-sign/src/main/java/org/ofdrw/sign/StandFormatAtomicSignID.java
@@ -13,12 +13,6 @@ import java.util.regex.Pattern;
 public class StandFormatAtomicSignID implements SignIDProvider {
 
     /**
-     * 根据标准推荐ID样式为
-     * <p>
-     * 'sNNN',NNN从1起。
-     */
-    public static final Pattern IDPattern = Pattern.compile("s(\\d{3})");
-    /**
      * 签名ID自增提供者
      */
     private final AtomicInteger provider;
@@ -33,7 +27,7 @@ public class StandFormatAtomicSignID implements SignIDProvider {
      * @param maxSignID 最大签名ID字符串
      */
     public StandFormatAtomicSignID(String maxSignID) {
-        int maxSignIDNum = parseIndex(maxSignID);
+        int maxSignIDNum = this.parse(maxSignID);
         provider = new AtomicInteger(maxSignIDNum);
     }
 
@@ -47,7 +41,7 @@ public class StandFormatAtomicSignID implements SignIDProvider {
      */
     @Override
     public void setCurrentMaxSignId(String maxSignId) {
-        int maxSignIDNum = parseIndex(maxSignId);
+        int maxSignIDNum = this.parse(maxSignId);
         provider.set(maxSignIDNum);
     }
 
@@ -80,23 +74,7 @@ public class StandFormatAtomicSignID implements SignIDProvider {
      */
     @Override
     public int parse(String id) {
-        return StandFormatAtomicSignID.parseIndex(id);
-    }
-
-    /**
-     * 解析出电子签名的ID数字
-     *
-     * @param id ID字符串
-     * @return ID数字
-     */
-    public static int parseIndex(String id) {
-        Matcher m = StandFormatAtomicSignID.IDPattern.matcher(id);
-        if (m.find()) {
-            String idNumStr = m.group(1);
-            return Integer.parseInt(idNumStr);
-        } else {
-            return Integer.parseInt(id);
-        }
+        return SignIdParser.parseIndex(id);
     }
 
 }