+++ ofdrw-sign/src/test/java/org/ofdrw/sign/stamppos/RidingStampPosTest.java
@@ -4,6 +4,7 @@ import org.junit.jupiter.api.Test;
 import org.ofdrw.gm.cert.PKCS12Tools;
 import org.ofdrw.gm.ses.v4.SESeal;
 import org.ofdrw.reader.OFDReader;
+import org.ofdrw.sign.NumberFormatAtomicSignID;
 import org.ofdrw.sign.OFDSigner;
 import org.ofdrw.sign.SignMode;
 import org.ofdrw.sign.signContainer.SESV4Container;
@@ -24,9 +25,8 @@ import static org.junit.jupiter.api.Assertions.*;
  * @author 权观宇
  * @since 2020-04-25 12:31:43
  */
-
-
 class RidingStampPosTest {
+
     @Test
     public void rideRight() throws GeneralSecurityException, IOException {
         Path userP12Path = Paths.get("src/test/resources", "USER.p12");
@@ -132,6 +132,62 @@ class RidingStampPosTest {
             signer.setSignContainer(signContainer);
             // 4. 设置显示位置
             signer.addApPos(new RidingStampPos(Side.Top, 40, 40));
+            // 5. 执行签名
+            signer.exeSign();
+            // 6. 关闭签名引擎，生成文档。
+        }
+        System.out.println(">> 生成文件位置: " + out.toAbsolutePath().toAbsolutePath());
+    }
+
+    @Test
+    public void rideRightOffset() throws GeneralSecurityException, IOException {
+        Path userP12Path = Paths.get("src/test/resources", "USER.p12");
+        Path sealPath = Paths.get("src/test/resources", "UserV4.esl");
+
+        PrivateKey prvKey = PKCS12Tools.ReadPrvKey(userP12Path, "private", "777777");
+        Certificate signCert = PKCS12Tools.ReadUserCert(userP12Path, "private", "777777");
+        SESeal seal = SESeal.getInstance(Files.readAllBytes(sealPath));
+
+        Path src = Paths.get("src/test/resources", "VPage7.ofd");
+        Path out = Paths.get("target/V4RideRightOffset.ofd");
+        // 1. 构造签名引擎
+        try (OFDReader reader = new OFDReader(src);
+             OFDSigner signer = new OFDSigner(reader, out, new NumberFormatAtomicSignID())) {
+            SESV4Container signContainer = new SESV4Container(prvKey, seal, signCert);
+            // 2. 设置签名模式
+            signer.setSignMode(SignMode.WholeProtected);
+            // 3. 设置签名使用的扩展签名容器
+            signer.setSignContainer(signContainer);
+            // 4. 设置显示位置
+            signer.addApPos(new RidingStampPos(Side.Right, 10d, 40, 40));
+            // 5. 执行签名
+            signer.exeSign();
+            // 6. 关闭签名引擎，生成文档。
+        }
+        System.out.println(">> 生成文件位置: " + out.toAbsolutePath().toAbsolutePath());
+    }
+
+    @Test
+    public void rideTopOffset() throws GeneralSecurityException, IOException {
+        Path userP12Path = Paths.get("src/test/resources", "USER.p12");
+        Path sealPath = Paths.get("src/test/resources", "UserV4.esl");
+
+        PrivateKey prvKey = PKCS12Tools.ReadPrvKey(userP12Path, "private", "777777");
+        Certificate signCert = PKCS12Tools.ReadUserCert(userP12Path, "private", "777777");
+        SESeal seal = SESeal.getInstance(Files.readAllBytes(sealPath));
+
+        Path src = Paths.get("src/test/resources", "VPage7.ofd");
+        Path out = Paths.get("target/V4RideTopOffset.ofd");
+        // 1. 构造签名引擎
+        try (OFDReader reader = new OFDReader(src);
+             OFDSigner signer = new OFDSigner(reader, out,new NumberFormatAtomicSignID())) {
+            SESV4Container signContainer = new SESV4Container(prvKey, seal, signCert);
+            // 2. 设置签名模式
+            signer.setSignMode(SignMode.WholeProtected);
+            // 3. 设置签名使用的扩展签名容器
+            signer.setSignContainer(signContainer);
+            // 4. 设置显示位置
+            signer.addApPos(new RidingStampPos(Side.Top, 10d, 40, 40));
             // 5. 执行签名
             signer.exeSign();
             // 6. 关闭签名引擎，生成文档。